Rename: 청년공간 카드/지도 토글 페이지 폴더 구분
