+++ src/app/(hasNotMap)/youth-place/(card)/page.module.scss
