+++ src/app/(hasNotMap)/youth-place/(card)/page.tsx
@@ -1,7 +1,7 @@
 import styles from './page.module.scss';
-import Filter from './_components/Filter';
-import ToggleButton from './_components/ToggleButton';
-import FilterablePlaces from './_components/FilterablePlaces';
+import Filter from '../_components/Filter';
+import ToggleButton from '../_components/ToggleButton';
+import FilterablePlaces from '../_components/FilterablePlaces';
 
 export default function PlacePage() {
   return (
