+++ src/app/(hasNotMap)/youth-place/(map)/placeMapList/page.module.scss
